Define helpers type in helper file
